+++ src/back_end/asm.rs
@@ -1,5 +1,3 @@
-// IDEA(stefano): split into nasm files and include them
-
 pub(crate) static CRASH_ASM: &str = {
     r"; fn ! = crash(msg: str @rdi:rsi, line: u64 @rdx, col: u64 @rcx)
 crash: